New files

+++ src/main/java/com/example/bankapplication/blik/BlikRequestDTO.java
@@ -2,16 +2,18 @@ package com.example.bankapplication.blik;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BlikRequestDTO {
 
 	private String blikCode;
 	private String requesterEmail;
-	private double requestedFunds;
+	private BigDecimal requestedFunds;
 	private String status;
 	private String message;
 
-	public BlikRequestDTO(String blikCode, String requesterEmail, double requestedFunds, String status, String message) {
+	public BlikRequestDTO(String blikCode, String requesterEmail, BigDecimal requestedFunds, String status, String message) {
 		this.blikCode = blikCode;
 		this.requesterEmail = requesterEmail;
 		this.requestedFunds = requestedFunds;
